+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 - Add `hostname` to `ionoscloud_server` resource and data source
 - Add `hostname` to `ionoscloud_vcpu_server` resource and data source
+- Add `hostname` to `ionoscloud_cube_server` resource and data source
 
 ## 6.6.0
 ### Refactor